feat: Termina filtro y busqueda e implementa navegacion a apartment_detail
