+++ Producto/renta_hogar/app/templates/customer/customer_apartment_detail.html
@@ -1,0 +1,20 @@
+{% extends "headers/home_header.html" %}
+
+{% load static %}
+
+{% block style_css %}
+
+<link rel="stylesheet" href="{% static 'css/home_header.css' %}">
+
+{% endblock %}
+
+{% block contenido %}
+
+<div class="content-login">
+
+    <h1 class="login-title">Aquí se mostraran los detalles del piso</h1>
+
+    
+</div>
+    
+    {% endblock %}
