+++ Producto/renta_hogar/app/templates/headers/home_header.html
@@ -12,7 +12,7 @@
     <header>
         <div class="navbar">
             <div class="logo">
-                <a href="{% url 'home' %}">  
+                <a href="{% url 'home_search' %}">  
                     <img src="{% static 'images/logo.png' %}" alt="Logo" style="height: 80px; margin-right: 10px;"> 
                 </a>                
             </div>
