+++ Producto/renta_hogar/app/templates/home.html
@@ -10,7 +10,7 @@
 {% block contenido %}
 <div class="contenido-home">
     <div class="filtros-container">
-        <form method="get" action="{% url 'home' %}" class="filtros" id="search_form">
+        <form method="get" action="{% url 'home_search' %}" class="filtros" id="search_form">
             <div class="filtro">
                 <label for="price_min">Precio mínimo/noche:</label>
                 <input type="number" id="price_min" name="precio_min" placeholder="Ej: 50"
@@ -33,57 +33,60 @@
                 <label for="fecha_fin">Fecha de salida:</label>
                 <input type="date" id="fecha_fin" name="fecha_fin" value="{{ request.GET.fecha_fin }}">
             </div>
-            <button type="reset" class="boton-borrar">Borrar filtros</button>
+            <button type="button" class="boton-borrar" id="clear_filters">Borrar filtros</button>
         </form>
     </div>
 
     <div class="resultados">
         <h2>Resultados de búsqueda</h2>
         {% if apartments %}
-        <div>
-            <a href="/auth/login"></a>
             <ul>
                 {% for apartment in apartments %}
-                <li>
-                    <img src="{{ apartment.photos.first.photo.url }}" alt="Foto de {{ apartment.address }}"
-                        class="apartment-photo">    
-                    <p class="apartment-details">Número de huéspedes: {{ apartment.guest_count }}</p>
-                    <p class="apartment-details">Dirección: {{ apartment.address }}</p>
-                    <p class="price">Precio/noche: {{ apartment.price }}€</p>
-                </li>   
+                <a href="/auth/login">
+                    <li>
+
+                        <img src="{{ apartment.photos.first.photo.url }}" alt="Foto de {{ apartment.address }}"
+                                class="apartment-photo">    
+                        <p class="apartment-details">Número de huéspedes: {{ apartment.guest_count }}</p>
+                        <p class="apartment-details">Dirección: {{ apartment.address }}</p>
+                        <p class="price">Precio/noche: {{ apartment.price }}€</p>
+                    </li>  
+                </a> 
                 {% endfor %}
             </ul>
-        </div>
         {% else %}
         <p>No se encontraron apartamentos que coincidan con los filtros.</p>
         {% endif %}
     </div>
 </div>
-{% endblock %}
+
 
 <script>
-    window.onload = function () {
-        const urlParams = new URLSearchParams(window.location.search);
-        if (!urlParams.toString()) {
-            console.log("Enviando formulario al cargar la página");
-            document.getElementById('search_form').submit();
+    document.addEventListener('DOMContentLoaded', function() {
+
+        const clearButton = document.getElementById('clear_filters');
+        if (clearButton) {
+            clearButton.addEventListener('click', function () {
+                console.log("Borrando filtros y recargando página...");
+                const baseUrl = window.location.origin + window.location.pathname;
+                window.location.href = baseUrl;  
+            });
         } else {
-            console.log("Parámetros detectados en la URL, no se envía automáticamente.");
+            console.log('El botón "Borrar filtros" no se ha encontrado.');
         }
-    };
-
-    const filtros = document.querySelectorAll('.filtro input');
-    filtros.forEach(filtro => {
-        filtro.addEventListener('change', function () {
-            console.log("Cambio detectado en:", filtro.name, "Valor:", filtro.value);
-            document.getElementById('search_form').submit();
-        });
-    });
-
-     document.getElementById('search_form').addEventListener('reset', function (event) {
-        setTimeout(() => {
-            console.log("Formulario reseteado. Enviando...");
-            document.getElementById('search_form').submit();
-        }, 0); 
+        const buscarButton = document.querySelector('.btn-buscar');
+        if (buscarButton) {
+            buscarButton.addEventListener('click', function (event) {
+                event.preventDefault(); // Prevenir el envío automático del formulario
+                console.log("Buscando con los filtros...");
+                document.getElementById('search_form').submit(); // Enviar el formulario manualmente
+            });
+        } else {
+            console.log('El botón "Buscar" no se ha encontrado.');
+        }
     });
 </script>
+{% endblock %}
+
+
+
