+++ Producto/renta_hogar/app/views/auth.py
@@ -71,9 +71,9 @@ def edit_profile(request):
 @login_required
 def delete_account(request):
     request.user.delete()
-    return redirect('home')
+    return redirect('home/search')
 
 @login_required
 def user_logout(request):
     logout(request)  
-    return redirect('home')  
+    return redirect('home/search')
